Moved examples code into get_light_curve.py
